[refactor]: API 쿼리 구조 변경 및 새로운 쿼리 키와 스키마 추가
